% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.R
\name{drop_dense}
\alias{drop_dense}
\title{Internal function to prune quantiles of non-important values for
visualization.}
\usage{
drop_dense(x, y, N_hard = 10000)
}
\arguments{
\item{x}{A numeric vector of sample/theoretical points.}

\item{y}{A numeric vector of theoretical/sample points.}

\item{N_hard}{Desired upperbound on number of points to plot.}
}
\value{
data.frame with o and e pruned as columns.
}
\description{
This function is not exposed, since we want to hard-code the parameters
for simplicity of usage.
}
