% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.R
\name{qq_drop_dense}
\alias{qq_drop_dense}
\title{Creates a Q-Q plot, with redundant points removed}
\usage{
qq_drop_dense(pvector, ...)
}
\arguments{
\item{pvector}{A numeric vector of p-values.}

\item{...}{Other arguments passed to \code{plot()}}
}
\description{
Creates a quantile-quantile plot from p-values from a GWAS study. We compare
the data quantile with a theoretical quantile from a uniform distribution.
This code is mostly adapted from the \code{qqman} package, but improved
for speed. Here we drop redundant points, where they are so dense in the plot
that we would not see them anyways. This makes the plot much faster.
}
\examples{
\dontrun{
qq_drop_dense(stats::runif(1e6))
}
}
\keyword{qq}
\keyword{qqplot}
\keyword{visualization}
