% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq.R
\name{qq}
\alias{qq}
\title{Creates a Q-Q plot}
\usage{
qq(pvector, ...)
}
\arguments{
\item{pvector}{A numeric vector of p-values.}

\item{...}{Other arguments passed to \code{plot()}}
}
\description{
Creates a quantile-quantile plot from p-values from a GWAS study. We compare
the data quantile with a theoretical quantile from a uniform distribution.
This code is mostly adapted from the \code{qqman} package, but improved
for speed. A graph with a hundred million points should only take a few
seconds to generate. Note that the graph is rasterised, so you should
specify the size of the plot in pixels, (that you plan to export), in the
input.
}
\examples{
\dontrun{
qq(stats::runif(1e6))
}
}
\keyword{qq}
\keyword{qqplot}
\keyword{visualization}
